vision/yolo_v4/pytorch/requirements.txt
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
fvcore==0.1.5.post20210727
matplotlib==3.3.4
numpy==1.19.5
pycocotools==2.0.4
pytest==6.2.5
pytest-pythonpath==0.7.4
requests==2.27.1
ruamel.yaml==0.17.21
scipy==1.5.4
torchvision==0.14.1+cpu; python_version > '3.6'
torchvision==0.11.0+cpu; python_version <= '3.6'
protobuf==3.19.4
wandb==0.12.8
yacs==0.1.8
git+https://github.com/graphcore/examples-utils@latest_stable
torchinfo==1.5.4
